% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrends.r
\name{cran_stats_by_day}
\alias{cran_stats_by_day}
\title{CRAN Stats for a specific date}
\usage{
cran_stats_by_day(Date, cran_log = "http://cran-logs.rstudio.com/")
}
\arguments{
\item{Date}{date of CRAN log to be downloaded}

\item{cran_log}{CRAN mirror name}
}
\value{
A dataframe with complete log data of a specific date
}
\description{
Download a full CRAN Log for a specific date
}
\author{
Avi Blinder
}

