\name{sample_log}
\alias{sample_log}
\docType{data}
\title{"Sample Formatted CRAN Log" }
\description{This sample formatted log is used as example in this package")}
\usage{data("sample_log")}
\format{
 date           : chr  "2016-11-01" "2016-11-01" "2016-11-01" "2016-11-01" ...

 time           : chr  "19:46:35" "19:47:38" "18:25:28" "18:29:46" ...

 size           : chr  "538958" "283733" "538964" "523" ...

 r_version      : chr  "3.3.2" "NA" "3.3.1" "3.3.1" ...

 r_arch         : chr  "x86_64" "NA" "x86_64" "x86_64" ...

 r_os           : chr  "darwin13.4.0" "NA" "mingw32" "darwin16.0.0" ...

 package_name   : chr  "sparklyr" "sparklyr" "sparklyr" "sparklyr" ...

 package_version: chr  "0.4" "0.4" "0.4" "0.4" ...

 country        : chr  "HK" "GB" "BR" "US" ...

 ip_id          : chr  "443" "659" "1117" "404" ...

}
\details{Partial CRAN log data}
\source{"http://cran-logs.rstudio.com"}
\examples{
data(sample_log)
}
\keyword{datasets}
