% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrends.r
\name{cran_stats_by_package}
\alias{cran_stats_by_package}
\title{CRAN Stats for single package}
\usage{
cran_stats_by_package(from_date, to_date, sel_package_name)
}
\arguments{
\item{from_date}{Starting date for downloading CRAN Repositories}

\item{to_date}{end date for downloading CRAN Repositories}

\item{sel_package_name}{Single package name for analysis}
}
\description{
Download CRAN data related to several packages in a range of dates
}
\examples{
cran_stats_by_package('2016-11-01','2016-11-02',"sparklyr")
}
\author{
Avi Blinder
}

