% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrends.r
\name{cran_stats_by_packages}
\alias{cran_stats_by_packages}
\title{CRAN Stats for several packages}
\usage{
cran_stats_by_packages(from_date, to_date, sel_package_name)
}
\arguments{
\item{from_date}{Starting date for downloading CRAN Repositories}

\item{to_date}{end date for downloading CRAN Repositories}

\item{sel_package_name}{Single package name for analysis}
}
\value{
dataframe with log information
}
\description{
Download CRAN log data related to several packages in
a range of dates
}
\author{
Avi Blinder
}

