% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrends_analysis.r
\name{cran_summary}
\alias{cran_summary}
\title{Merge stats with County names}
\usage{
cran_summary(cran_log)
}
\arguments{
\item{cran_log}{dataftame containing CRAN log data
data(sample_log)
cran_summary(sample_log)}
}
\description{
Merge CRAN log data with County names
}
\author{
Avi Blinder
}

